requirements.txt
BeautifulSoup==3.2.1
Django==1.9.4
django-braces==1.8.1
django-datatable==0.2.0
django-model-report==0.2.1
django-mssql==1.7
django-pure-pagination==0.3.0
django-pyodbc==0.3.0
django-pyodbc-azure==1.9.3.0
django-spreadsheetresponsemixin==0.1.10
django-tables2==1.1.6
django-tables2-reports==0.0.10
django-wkhtmltopdf==3.0.0
djangorestframework==3.3.2
et-xmlfile==1.0.1
html5lib==0.9999999
jdcal==1.2
openpyxl==2.3.5
Pillow==3.1.0
Pygments==2.1
pyodbc==3.0.10
PyPDF2==1.25.1
reportlab==3.3.0
six==1.10.0
South==1.0.2
sql-server.pyodbc==1.0
tablib==0.10.0
xhtml2pdf==0.0.6
xlwt==0.7.4
